+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Version History
 ### Changes in v1.2.0:
 
 -   Slightly improved denoising speed
+-   Link C/C++ runtime library dynamically by default on Windows too
+-   Added OIDN_STATIC_RUNTIME CMake option (for Windows only)
 
 ### Changes in v1.1.0: