+++ CHANGELOG.md
@@ -10,8 +10,10 @@ Version History
 -   Improved denoising speed by about 7-38% (mostly depending on the compiler)
 -   Link C/C++ runtime library dynamically on Windows too by default
 -   Added `OIDN_STATIC_RUNTIME` CMake option (for Windows only)
--   Added support for OpenImageIO to the examples (disabled by default)
+-   Added support for OpenImageIO to the example apps (disabled by default)
 -   Added check for minimum supported TBB version
+-   Find debug versions of TBB
+-   Added testing
 
 ### Changes in v1.1.0: