+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version History
 -   Link C/C++ runtime library dynamically on Windows too by default
 -   Renamed example apps (`oidnDenoise`, `oidnTest`)
 -   Added benchmark app (`oidnBench`)
+-   Fixed compile error for ISPC 1.13+
 -   Fixed training warning with PyTorch 1.5+
 
 ### Changes in v1.2.0: